refactor(core): moved special tests providers to new folder
