+++ src/modules/test.module.ts
@@ -12,17 +12,17 @@ import { SectionProvider } from '../providers/section.provider';
 import { UserModule } from './user.module';
 import { SimpleLightTestEntity } from '../entities/simple-light-test.entity';
 import { SimpleLightTestController } from '../controllers/simple-light-test.controller';
-import { SimpleLightTestProvider } from '../providers/simple-light-test.provider';
+import { SimpleLightTestProvider } from '../providers/tests-providers/simple-light-test.provider';
 import { TestValidationStrategy } from '../strategies/test-validation.strategy';
 import { SimpleSoundTestController } from '../controllers/simple-sound-test.controller';
-import { SimpleSoundTestProvider } from '../providers/simple-sound-test.provider';
+import { SimpleSoundTestProvider } from '../providers/tests-providers/simple-sound-test.provider';
 import { SimpleSoundTestEntity } from '../entities/simple-sound-test.entity';
 import { HardLightTestEntity } from '../entities/hard-light-test.entity';
 import { HardLightTestController } from '../controllers/hard-light-test.controller';
-import { HardLightTestProvider } from '../providers/hard-light-test.provider';
+import { HardLightTestProvider } from '../providers/tests-providers/hard-light-test.provider';
 import { AdditionTestEntity } from '../entities/addition-test.entity';
 import { AdditionTestController } from '../controllers/additional-test.controller';
-import { AdditionTestProvider } from '../providers/additioon-test.provider';
+import { AdditionTestProvider } from '../providers/tests-providers/additioon-test.provider';
 import { JwtTestLinksGeneratorUtil } from '../utils/jwt-test-links-generator.util';
 import { JwtGuardModule } from './jwt.guard.module';
